+++ src/FolkerKinzel.VCards/md/ReleaseNotes.md
@@ -7,7 +7,8 @@ IAsyncEnumerable<VCard> DeserializeManyAsync(IEnumerable<Func<CancellationToken,
                                              [EnumeratorCancellation] CancellationToken)
 ```
 is available now for all frameworks the package supports.
-- The new option `Opts. WriteRfc2739Extensions` is set as default and allows to write the properties `FBURL`, `CALURI`, and `CALADRURI` in vCard 3.0.
+- The new option `Opts. WriteRfc2739Extensions` is set as default and allows to write the properties `FBURL`, `CALURI`, `CAPURI`, and `CALADRURI` in vCard 3.0.
+- The VCard class and the VCardBuilder class got a new `CalendarAccessUris` property, which supports the RFC 2739 `CAPURI` extension in vCard 3.0.
 - Dependency updates
 - High code coverage
 &nbsp;